+++ counter_slider/CHANGELOG.md
@@ -4,7 +4,7 @@
  - Fixed documentation. #TODO
  - Renamed `setValue` to `onChanged`.
  - Renamed `buttonBorderGap` to `gapSize`
- - Change fixed width + height to optional, default unconstrained. #TODO
+ - Change fixed width + height to optional, default unconstrained.  
  - Fixed `borderSize: 0` still visible.
 
 ## 0.0.2+1